Fixes Mantrap cheese abuse (#1288)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Makes it so Mantraps are too large to fit in anywhere, while armed. If
you want to fight with em, learn to place ahead of time, and lure people
into them.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Some of the mantrap cheese is down right abusive, and bad for gameplay.
This should no longer happen. As an example of said bad play see a small
bit of the following:


https://github.com/user-attachments/assets/6339b3be-dac5-48ca-b79a-ec6545f15ae1


<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->