Fixes an ancient admin-tooling gripe I've had (#1312)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
adds a preference to make you not have the super special blue text as an
admin speaking in ooc
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
admins don't need to fully deadmin every single time they just want to
talk like a normal player
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Co-authored-by: zoktiik <zoktiik@users.noreply.github.com>