Changes how to put someone in a Pillory (#1326)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Pillories now require you to have someone in an aggressive grab to be
able to shove them into one.
(Had to make due with checking all grabs since it doesn't ask for your
mob)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
It made combat easy to cheese if you where near one, since you can
instantly put someone in a Pillory **even if they are not next to it**.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->