Ruralization - Two Cottages (#1329)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Two lovely cottages fit for couples have been built outside Rockhill!


![image](https://github.com/user-attachments/assets/cfb67cd6-059f-4879-ae61-81a32470d04b)

### Here, we have the Farmer Cottage:


![image](https://github.com/user-attachments/assets/de914995-f08b-4065-a995-3786112c3df4)

As you can see, fair space for lovemaking and the cooking of homemade
meals.


![image](https://github.com/user-attachments/assets/8f0519a6-e476-4ac5-a400-ba8d68937dd0)

Outside you may find a small patch of land to be plowed, a hoe has been
stocked inside the houses wardrobe.

### Here, we have the Woodman Cottage:


![image](https://github.com/user-attachments/assets/98c033ba-20ba-4c9e-9912-fe404896d955)

While it is small, you have space in the back for all your lumbering
needs!


![image](https://github.com/user-attachments/assets/22c8c277-5b5b-4e2f-810f-28bcc190df6f)

However, the house includes a second floor, where the bedroom is.


![image](https://github.com/user-attachments/assets/66591235-2882-4d1a-86f1-baa5b0117e23)

And for all you lovebirds, we have included a balcony fit for a romantic
dinner...


![image](https://github.com/user-attachments/assets/04c5eea6-6c89-4465-9249-a04762bf6471)

We hope you enjoy the rural side of Rockhill your future children will
thank you for it!

## Why It's Good For The Game

There are two reasons for doing this:
1 - If everything goes well, space will be required in the city to
accommodate for more interesting shops and attractions.
Creating some houses outside of the city allows for houses within the
city to be repurposed later without losing anything in the process.
2 - There was a debate in favor of Deurbanization in discord's feedback
channel.
(Here's Wow malding on the topic because its funny)

![image](https://github.com/user-attachments/assets/8f41181b-3fcc-49e7-a132-74a788836542)
These houses I created accommodate players who may want to RP a low
class medieval family life. Something the houses inside the city are
just not inviting people to do.
They are modeled after the Farmer and Woodcutter Towner roles.

If this is well accepted I plan to expand and with time build a little
village there. But in parts, to simulate the construction of it as if it
was taking time like in real life.
I already added the well there to be the "neighborhoods" source of water
to ensure the peasants there can subside.

This also hopefully brings some dynamic and light to a part of the map
that was left unused. (Darkness, no trees, it was like it was WAITING to
have something built on it!)