Lantern for Miner (#1333)
## About The Pull Request

At the request of a few people, the miner now has a lantern. It makes
sense, given they're meant to be underground a majority of the time.

## Why It's Good For The Game

Lanterns should've already existed for Miners. It doesn't really make
sense that they'd be swinging a pick and holding a torch at the same
time.