Ranged seelie nerf (#1339)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Reduces seelie ranged damage to be in line with what their melee damage
should be (10% due to 1 str)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Seelies are not supposed to engage in combat and if they are they are
meant to do very little damage, nerfing the ranged damage nerfs the crit
rate for seelies using bows and crossbows in line with what their melee
damage is.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->