Harder Smithing for the Unskilled (#1340)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This was made as a result of various discussions of this topic on
discord.


![image](https://github.com/user-attachments/assets/6a7d87fc-6cbb-4dbe-8b22-b4497a5b0562)
***Funny thing by Pom***

This makes i much harder to get into smithing if you have 0
blacksmithing skill.

If you have 0 smithing skill you will have a 50% to not make any
progress while smithing.
You then have a 25% chance to ruin the bar.

Old values were 25% and 12% respectively.

This makes it more costly to learn smithing. While not making it
impossible.

## Why It's Good For The Game

There are 4 smithing classes. Guildsmith, Smith Apprentice, Towner
Blacksmith, and Forge Cleric.
If players want to smith, they should choose one of these roles.
Choosing a role should be a compromise, you're getting certain skills in
exchange for not getting others.
There isn't a main character, and everyone is necessary for the town to
work.
This will also hopefully reduce smithery break-ins.
Goodfaith players will not suffer form this change.

## Changelog

balance: Decreased the smithing success chance for characters who have
no smithing skill.