Physician can now be a graggarite, similar to town doctor (#1365)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
physician can now be a graggarite, because physician is a commoner,
theres no real reason a graggarite couldnt pose as a pestran in order to
get to a position where they can silently take a bite out of nobles to
serve their own desires

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

physician can now be graggarite, similar to the town doctor

(yumby, human met c: )

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->