Female Barbarians! (#1373)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This PR simply lets female characters spawn in as a Barbarian.

## Why It's Good For The Game

There's not really a reason for female barbarians to not be a thing. I
don't want to say it's chudcode, but I feel like the archetype of Female
Barbarian is different enough in both gameplay and fantasy from Amazon
that it warrants a change. Especially since some players (me,
admittedly) would like to embody that archetype in a slightly more
serious way than a buff girl in a chainmail bikini.