Many Changes, Detailed Below
Altered global define for Seelie, and two new lists to help include them in jobs such as Hand, Servant, and Seamster.
Allowed Seelie to become hand and servant, with custom flavor text on starting, as well as custom flavor text for starting as a prisoner.
Reworked water bolt spell, rebalance of cooldown delay on most Seelie spells
New additions of spells, such as the AOE slowdown spell for Seelie Hand and the Sate Crop spell for Seelie Maid.