Allows emoting wile inside a thing. (#1384)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Removes a check that made emotes blank wile inside chests, closets,
handcarts and lampterns.
(played around a bit with it and doesn't seem to have broken anything by
removing it)


![lamptern](https://github.com/user-attachments/assets/f05182fc-9703-4085-90f1-0709839d58f5)

![cart](https://github.com/user-attachments/assets/e0cdf180-f0b7-4ba0-9b35-56af361a1013)

![closet](https://github.com/user-attachments/assets/27b2d353-d893-41a6-afb8-243806a022af)

![chest](https://github.com/user-attachments/assets/d663b504-b20d-4007-95dc-665a11fba7f7)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Heard a few people that wanted this so I decide it to "fix" it.
(though it seems it was an intentional thing)
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->