[TM] Ratwood goes Punk (#1385)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Replaces the majority of torches and braziers that served as road lights
with actual road lights powered by an unknown energy.
The city will have a more grim and medievalpunk tone to it, and the many
minor alleyways will be less evenly lighted.


![image](https://github.com/user-attachments/assets/655ab2a4-32af-4c14-8536-aa3c829f9596)

![image](https://github.com/user-attachments/assets/032acf22-26bc-4448-aa7d-2130ad4455eb)

![image](https://github.com/user-attachments/assets/a2112250-3f95-4c3e-8268-75ab4069429d)

![image](https://github.com/user-attachments/assets/464c7aff-8bc0-4211-a96d-b79d6249a37c)


## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

The Medieval age is coming to a close.