Remove explicit mentions of rape and r slurs (#1388)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Title. The hate we received when we first opened got our repo flagged.
We are gonna have to open a support ticket for it but best to look good
when that happens.

Also, 

> 3 - Do not include slurs in your code, comments included.