Mercenary Guild QoL
Remaps the entirety of the mercenary's guild to be much more appropriate for what it is. Has an office for the to-be-mentioned Captain, a dining hall and meeting area.

Adds the Captain for the Grenzelhoft mercenaries as a class. Single slot, of course.
Additionally, it slaps in a clerk class, functioning as you'd expect. Cooking, cleaning and writing contracts.