Runelocks (#1364)
## About The Pull Request
Just pulls https://github.com/Tree445/Hearthstone/pull/218 and modifies
it for our use, alongside fixing it up some. Runed firearms, as opposed
to matchlocks. More fantasy feeling, IMO.

The idea is to give three shots via the ammo pouch to the Lord
Inquisitor, with an unloaded runelock pistol roundstart. They're
reusable projectiles, same as arrows and bolts, but are
guaranteed(mostly) to embed and may not fall out until later, rendering
the weapon useless in the meantime. Totally pointless to keep and haul
around if you lose the projectiles, too, since you can't get more.
Additionally, they provide an identical effect to werewolves that the
silver dagger does, should it hit them while transformed.

You get one shot, a lengthy reload, and not much to counter someone
rushing you beyond hitting them with a golf ball sized projectile. Try
not to miss.

For this, since the Inquisitor spawns with some additional equipment,
they now spawn with a backpack, and their bank account was updated due
to their lack of a pouch now.
## Why It's Good For The Game
It's not an arquebus, and it certainly isn't the club's special musket,
which I'd have preferred over this, but it's something at the least.
Variety to ranged combat would be good, even if this is so limited in
number.