New Cultist restrictions (#1392)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
blacklists roles that can't take zizo as a patron, from being cultists
round start.
Adds the following roles to the blacklist:
	"Bandit",
	"Knight",
	"Retinue Captain",
	"Gatemaster",
	"Bog Master",
	"Inquisitor",
	"Confessor",
	"Acolyte"
Notably, removes merchant from blacklist role. The merchant absolutely
can and should be able to be a cultist.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Roles that can't BE zizoites, because they are thoroughly vetted, should
not be starting as zizo cultists. I mean come on now. Having the guard
captain and knights be roundstart cultists and offering the royal family
40 minutes into the round is pain.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->