Magic balance pass/spell list changes (#1205)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR Re-balances a lot of spells in the mage spell list. Currently
the most common ones people ever take is fireball, lightning bolt, and
repulse for combat reasons. Some of the other ones such as Arcane bolt,
spitfire were some of the most useless things I've ever seen, and either
needed improvement or spellpoint decreases. This PR aims to rebalance
the spell list for things to be worth their spell cost, or reduced in
cost.
From the top:

**- Fireball:** Reduced in cost from 4 to 3. Fireball used to be the
best spell bar none, because of the way firestacks did damage. With how
firestacks were nerfed, unless someone is stupid enough to stand on the
fire after being hit by the fireball, you get anywhere from 20 to 60ish
damage or so on use.

**-Lightning bolt:** Will now shock everyone equally. Apparently, gloves
used to completely negate this from doing anything, ever. Gloves have
lost their anti-lightning magic

**-Spitfire:** Reworked. Formerly a worthless damage spell with 'fast'
cool down, It's now a burst fired attack. Upon activation, it will fire
a burst of three fire bolts that deal 15 burn damage each, for a total
of 45 damage, assuming all hit- a fairly big if. This spell has minor
homing features, but in likely hood you won't notice unless you're
running away from it, given the speed of the projectiles. Still fairly
easy to dodge, however if moving perpendicular to caster.

**-Arcane bolt:** Bumped up to 45 brute damage, and cost increased to 3.
This spell is the equivalent of shooting an arrow at someone if you're
decent at using a bow and arrow. Does have slight armor piercing.

**-Forcewall:** Formerly a noob trap spell, This was more or less only
used by people to trap themselves in with someone else. It has been
changed to allow the caster to walk through it, at the cost of reducing
the wall's existance to 10 seconds instead of 20.

**-Ensare:** A hold spell that was effectively worthless due to the
short time it held people in place. It was not long enough to cast even
the shortest castable spell on our spell list, leaving no way to take
advantage of holding someone, aside from fleeing. immobalize Duration
buffed to 4 seconds from 2.5

**-Bladeburst:** This spell was probably the hardest one to figure out
what to do with. It's atrocious. Extremely difficult to land a hit with,
and the damage it did spread all over the body, making it effectively
worthless in causing any significant harm on someone. I reworked this
spell to instead direct all that damage to a randomly picked leg, and
fracture a leg bone. This might be an over compensation, but it's hard
to tell, and as I said, it's extremely hard to even land a hit with
this. Might need a point increase to 3, we'll have to see.

**-Haste:** A interesting spell, It's duration was increased to 1.5
minutes, and cooldown of casting it halfed to 2.5 minutes from 5. It's
cost was reduced to 2 spell points from 3, and it recieved near double
the stamina to cast. Better now for use to buff others, though you can
still buff yourself mid fight in a pinch.


![image](https://github.com/user-attachments/assets/4df20a32-d9df-4d0b-875f-869fc91e3fd1)

![image](https://github.com/user-attachments/assets/ed457587-9f90-4ba0-9414-0c7f85ec7694)

![image](https://github.com/user-attachments/assets/411813c4-2028-47ac-9bbc-a4c0a38ec85d)


----

This PR also adds a new spell that I couldn't resist adding- the
traditional find familiar spell. Spawns a spectral looking volf that is
hostile to everyone except the caster, and disappears after 30 seconds.

---
Finally, This PR adjusts arcane spell XP gain. In normal gameplay as a
mage, you would /never/ level up, even as a novice, unless you
intentionally grinded spellcasting at a wall or trees, or-anything. As
an example for how bad this was, It would take /50/ fire balls to
advance from apprentice, to journeyman. Further more due to the way XP
was calculated based on fatigue spent, the higher INT level you were as
a mage, the less fatigue you used, meaning the better mage you were, the
worse you were at learning. Thus:
The overall spell xp multiplier has been increased from .3 to .5, and a
multiplier has been added based on the mage's INT level, divided by ten.
This functions the same way a archers bow damage is increased.
EX: an 11 int mage will gain 1.1 times the magic xp they would have.
Likewise, a 15INT mage gains 1.5 xps the xp.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Currently the most common ones people ever take is fireball, lightning
bolt, and repulse for combat reasons. Some of the other ones such as
Arcane bolt, spitfire were some of the most useless things I've ever
seen, and either needed improvement or spell point decreases. It was
very clear that some spells were just horrendiously bad with no
redeeming qualities. Now, all the spells should be relatively more
inline with each other, letting people feel like they have more of a
choice with what to pick and use.

Also, No one should have to spam cast 50 spells to level up from
apprentice to journeyman. It gets much worse the higher the level, so
the xp gain had to be increased.

I'm saving most of the other spells I intend on adding at some point for
another PR: This one is predominantly for rebalancing existing ones.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->