Bounty (#1371)
This is a list of changes covering the Medical part of the bounty,
including a large variety of changes. Please note that because of the
size and novelty of the PR, bugs and adjustments are bound to happen.
Entire systems had to be written from scratch due to the severe lack of
them, and many work arounds had to be found to work with the current
code. Feedback is very appreciated, but keep this in mind when
criticizing this work.

This is not an extensive and exact list of the changes and recipes, and
is more meant to give you an idea of what has changed.

---------

Summary:

-Reworked the surgeon roles entirely, giving them new abilities and
items to work with.

-Many new craftable items.

-A new alchemy system. (Much of the alchemy rework is in preparation for
the second half of the bounty. Many changes are planned still, as this
is more of a base to work from.)

-A large variety of new medicines and drugs you can make or buy through
shops, mainly through the Purity vendor.

-Medicine and Alchemy were merged together into a new skill called
Treatment. (This is in due part in preparation for the plague and
disease part of the bounty. While some may see this as an
oversimplification, the result will be the opposite, as it will require
much more knowledge and skill by the surgeons as they will manage all
aspects of medicine now, allowing for more engagement especially when
patients are unavailable.)

-Added a new clinic next to the keep, to the right.

-New roles: Barber surgeon got renamed to Local Physician. Roles meant
solely for alchemy are removed due to their redundancy. Added new clinic
specific roles under yeomen (for now), Clinic Doc and Clinic Guard.

---------

New items:

-A new craftable rucksack, a downgraded backpack essentially. It comes
with 4 slots less than a backpack, but can be made with bag and rope.

-Improvised craftable surgery tools made out of rock, sticks and wood. A
cheap way to obtain surgery tools when none are available.

-An alembic and a new mortar and pestle. Used in alchemy, the mortar and
pestle are not considered 1 item instead of two separate items to craft,
and the alembic is used to take powdered essences from the mortar and
create chemical reactions. For example, to make a health potion, you
need to first powder and then run meat and poison berries through an
alembic.

-As above, a large variety of new powders and chemical reactions were
added to compliment the new system.

-A personal patch kit with a variety of medicines included.

-New obtainable medicines and drugs that can be gained from the Purity
vendor. These include syrettes, bandages and bottles with different
effects:

Detox; Purges chemicals and toxicity from the system.

Rejuv; Works like and EpiPen in base ss13. It's meant to stabilize dying
people in a pinch, giving you enough time to get them treated properly
without them dying on you.

PNKballs; Works like a weaker health potion, with the addition of
healing blood loss.

Wake-Up; A bag of 3 pills, used to starve away the need to sleep.
Restores fatigue and stores skillpoints for another day. Good if
sleeping is just not in your schedule for the day, without suffering
from the tiredness penalties.

Bandages; Come in a first aid kit. Can be crafted with 3 cloth and 1
ash. 1/3rd better than normal bandages with 3 uses.

---------

New surgery actions and abilities:

(Note; While these are technically spells in the code, this is due to
the sake of convenience rather then how it should be seen as literally:
This way, these abilities can be limited to the select specialized
medical roles, while also allowing surgeons to quickly mend critical
situations more effectively while helping with role identity. To add,
these abilities do not remove the need for surgery, but are actually
meant to complement the surgery system, and for many, are actually
intertwined with the surgery system and cannot be used without the
other.)

All of these abilities are TOUCH targeted, so you will need to be next
to your patient to use them.

-Rapid Mending: Functions similar to a lesser miracle, but with a much
longer cooldown. Useful for speeding up surgeries, but cannot be solely
depended upon.

-Purge: This ability purges Toxins and Chems from the body at the cost
of opening an artery. It has a very long cooldown, and is most useful
for purging chemicals for revival, or in an emergency. Due to it's
damaging ability, works best with additional drugs or Rapid Mending.
(Note: Some may see this and think it can be used in combat but as a
reminder, not only is the cooldown exceedingly long for it, but you
cannot use it if your target is moving, so it's very unlikely to be used
at all. Even so, a surgeon has no way to follow up on such an attack
given their lack of combat skills, and even if they were trying to
specialize in combat as a surgeon, this would be likely deemed as going
out of your intended role in the rules of the server.)

-Stabilize: Works like an EpiPen, stabilizes patients in critical
condition and cures all oxygen damage. Recommended before committing to
a heart transplant.

-PURITY: Helps seal wounds on the target, and if they are infected with
either zombie or werewolf bites, it will clear them in return for some
burn damage.

-Tissue Debridement: An ability that de-rots the target. Requires the
liver to be surgically modified for it, but can only be cast once per
liver, requiring a transplant for new casts. Cannot be cast otherwise.
Requires a bed.

-CPR: An ability that revives the target. Just like the above, requires
a surgery on the heart, modifying it. Will revive once the condition is
met, but to be revived again afterwards requires a new heart transplant.
Requires a bed.


this description was wrote by someone with better people skills than me,
my own touch, its a bounty, I get payed for this, dont muck about in the
code and do balance shit. PR's for bug fix's are fine. also dont @ me in
discord over this shit do another staff member.

---------

Co-authored-by: Cyprex <35031555+Cyprex@users.noreply.github.com>