Fetch spell can catch items (#1382)
## About The Pull Request

- Ports
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/696

> Really simple: if you have CATCH mode enabled while casting Fetch at
an item, you will attempt to catch whatever you're trying to fetch.
Note: throw_at is not perfect, especially on diagonals.
> 
> If you cannot catch it, it will be thrown at you directly instead. Be
smart.

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

It's a cool and niche mechanic to an otherwise boring spell.

Co-authored-by: Ephemeralis <Ephemeralis@users.noreply.github.com>