New color and bigger message for personal mail notification (#1396)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Makes the personal mail a different color, and slightly bigger

## Why It's Good For The Game


![image](https://github.com/user-attachments/assets/d8a54848-b013-4fe4-a65a-f537a589019b)
Current one

![image](https://github.com/user-attachments/assets/5aeb4829-d8f6-4ac5-a684-72a5be6127db)
After changes

Many people are unaware that you can send emails directly to people to
pick them up, and even if you do, the message is easy to miss. This PR
aims to make the notification more noticeable and thus harder to miss.