Changes how the title of the Lord is stored for announcements (#1399)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Now it's all stored in a global var that is already inizialiced with
"Duke" in case any of the announcements are called outside of the
Throne.
This var will now only change when the priest coronates a new ruler.


![Duchess](https://github.com/user-attachments/assets/acc7295e-fcaa-4929-ad2d-2cfd7701313c)

![Duke](https://github.com/user-attachments/assets/85a9a51d-9b7a-41d0-80eb-8594127c6dd9)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Fixes an issue introduced with #1375, that caused the outlaw special
trait to make an announcement with a wrong title.


![image](https://github.com/user-attachments/assets/3dc06ae9-dddf-4cdd-b94a-57d059846529)
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->