removed cant sleep from all but chest armor (#1395)
told some people when I finished the med rework I would do this. I
disagree with the armor sleep shit entirely its an extra un-needed step.
but I get its application for plate and heavy chest shit. this removes
the flag from all but chest armor. gloves, hats etc. IF you are wearing
plate shit odds are you ahve a chest piece for it since this is most
important. it makes it where only removing that is needed basically.
this trims the bullshit without removing the principle behind it.