Allow admins to toggle aghost invisibility (#1423)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

- Ports
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/759

proof of testing:

![image](https://github.com/user-attachments/assets/e82b3d1c-438c-4458-9315-589ef0e11448)


## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Not every admin wants their character to disappear when they aghost.
This also prevents the annoying bug where the character simply
disappears into the void and can no longer be gotten back into.

Co-authored-by: Ephemeralis <Ephemeralis@users.noreply.github.com>